# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vue Laravel Template API",
        "description": "API documentation for the Vue Laravel Template project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/blogs/{id}/react": {
            "post": {
                "tags": [
                    "Reactions"
                ],
                "summary": "React to a blog post",
                "operationId": "9a2a0cee8aca9b16b67d0e7318f85fa7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reaction_type"
                                ],
                                "properties": {
                                    "reaction_type": {
                                        "type": "string",
                                        "example": "like",
                                        "enum": [
                                            "like",
                                            "love",
                                            "haha",
                                            "wow",
                                            "sad",
                                            "angry"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reaction added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReactionResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Reaction removed or swapped",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReactionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ReactionResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Reaction registered"
                    },
                    "status": {
                        "type": "string",
                        "example": "attached"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}